+++ src/state/use-beta-account.js
@@ -5,8 +5,44 @@ import { usePushID } from "../utils/use-pushid";
 import { BetaAccountContext } from "./with-beta-account";
 
 const CREATE_ACCOUNT = gql`
-  mutation CreateAccount($uuid: String!, $data: jsonb = "{}") {
-    account: insert_beta_accounts_one(object: { uuid: $uuid, data: $data }) {
+  mutation CreateAccount($accountID: String!, $data: jsonb = "{}") {
+    account: insert_beta_accounts_one(
+      object: { uuid: $accountID, data: $data }
+    ) {
+      uuid
+    }
+  }
+`;
+
+const CREATE_PROJECT = gql`
+  mutation CreateDummyProject(
+    $accountID: String!
+    $projectID: String!
+    $title: String!
+    $data: jsonb!
+  ) {
+    project: insert_beta_projects_one(
+      object: {
+        uuid: $projectID
+        account_uuid: $accountID
+        title: $title
+        data: $data
+      }
+    ) {
+      projectID: uuid
+      created_at
+      updated_at
+      title
+    }
+  }
+`;
+
+const APPEND_PROJECT = gql`
+  mutation AppendDummyProject($accountID: String!, $projectID: jsonb!) {
+    update_beta_accounts_by_pk(
+      pk_columns: { uuid: $accountID }
+      _append: { own_projects: $projectID }
+    ) {
       uuid
     }
   }
@@ -15,15 +51,70 @@ const CREATE_ACCOUNT = gql`
 export const useBetaAccount = () => {
   const { uname } = useParams();
   const { generatePushID } = usePushID();
-  const [createFn] = useMutation(CREATE_ACCOUNT);
+  const [createAccountFn] = useMutation(CREATE_ACCOUNT);
+  const [createProjectFn] = useMutation(CREATE_PROJECT);
+  const [appendProjectFn] = useMutation(APPEND_PROJECT);
   const { setAccountID, loadAccount, ...state } =
     useContext(BetaAccountContext);
 
-  const createAccount = () => {
-    const uuid = generatePushID();
-    createFn({ variables: { uuid } }).then((res) =>
-      setAccountID(res.data.account.uuid)
-    );
+  const createAccount = async () => {
+    const accountID = generatePushID();
+    const projectID = generatePushID();
+    const t1 = generatePushID();
+    const t2 = generatePushID();
+
+    await createAccountFn({ variables: { accountID } });
+    await createProjectFn({
+      variables: {
+        accountID,
+        projectID,
+        title: "Cook Dinner",
+        data: {
+          collapse: [t2],
+          items: [
+            {
+              id: t1,
+              title: "Main coure is done"
+            },
+            {
+              id: generatePushID(),
+              parentId: t1,
+              title: "Boil water",
+              status: true
+            },
+            {
+              id: generatePushID(),
+              parentId: t1,
+              title: "Put pasta into boiling water"
+            },
+            {
+              id: generatePushID(),
+              parentId: t1,
+              title: "Serve with garlic and oil"
+            },
+            {
+              id: t2,
+              title: "Desserts are done"
+            },
+            {
+              id: generatePushID(),
+              parentId: t2,
+              title: "Buy a frozen cake",
+              status: true
+            },
+            {
+              id: generatePushID(),
+              parentId: t2,
+              title: "Move it to the fride before cooking time",
+              status: true
+            }
+          ]
+        }
+      }
+    });
+    await appendProjectFn({ variables: { accountID, projectID } });
+
+    setAccountID(accountID);
   };
 
   const redeemAccount = (accountId) =>